Fix crashes (#316)
* Fix NPE on inspecting non-existent files

FileStore#packageName and FileStore#modified used to raise NPE when a
specified file does not exist. This patch makes them return null so that
callers can handle them.

* Fix leaks of ReusableCompiler.Borrow

ReusableCompiler.Borrow used to be leaked when a runtime exception is
thrown in CompileBatch.needsAdditionalSources. This patch ensures to
release it. It should fix "Compiler is already in-use!" errors that
repeat after a runtime exception.